Refactored cpu_quota_manager
