+++ ydb/core/fq/libs/compute/common/ya.make
@@ -1,7 +1,6 @@
 LIBRARY()
 
 SRCS(
-    cpu_quota_manager.cpp
     pinger.cpp
     run_actor_params.cpp
     utils.cpp
